requirements.txt
absl-py==1.4.0
APScheduler==3.9.1
asttokens==2.2.1
astunparse==1.6.3
attrs==22.1.0
Automat==22.10.0
backcall==0.2.0
bookstore==0.0.1
cachetools==5.3.0
certifi==2022.9.24
cffi==1.15.1
chardet==5.1.0
charset-normalizer==2.1.1
click==8.1.3
cloudpickle==2.2.1
colorama==0.4.5
comm==0.1.2
constantly==15.1.0
contourpy==1.0.6
cryptography==38.0.4
cssselect==1.2.0
cycler==0.11.0
debugpy==1.6.6
decorator==5.1.1
distlib==0.3.6
dm-tree==0.1.8
dnspython==2.2.1
docopt==0.6.2
exceptiongroup==1.0.0rc9
executing==1.2.0
filelock==3.9.0
fixture==1.5.11
Flask==2.2.2
flatbuffers==23.3.3
fonttools==4.38.0
gast==0.4.0
gensim==4.3.1
google-auth==2.16.2
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==2.0.1
grpcio==1.51.3
h5py==3.8.0
hyperlink==21.0.0
identify==2.5.9
idna==3.4
ImageHash==4.3.1
incremental==22.10.0
iniconfig==1.1.1
ipykernel==6.21.3
ipython==8.11.0
itemadapter==0.7.0
itemloaders==1.0.6
itsdangerous==2.1.2
jedi==0.18.2
jieba==0.42.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
jsonpickle==3.0.1
jupyter_client==8.0.3
jupyter_core==5.2.0
jwt==1.3.1
keras==2.11.0
kiwisolver==1.4.4
libclang==15.0.6.1
llvmlite==0.39.1
lxml==4.9.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
miditok==2.0.0
miditoolkit==0.1.16
mido==1.2.10
more-itertools==9.1.0
mpmath==1.3.0
music21==8.1.0
nest-asyncio==1.5.6
nltk==3.8.1
nodeenv==1.7.0
numpy==1.23.3
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==21.3
palettable==3.3.0
pandas==1.5.1
parsel==1.7.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.2.0
pipreqs==0.4.13
platformdirs==2.5.4
pluggy==1.0.0
pretty-midi==0.2.10
prompt-toolkit==3.0.38
Protego==0.2.1
protobuf==3.19.6
psutil==5.9.4
psycopg2==2.9.5
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.1
pycparser==2.21
PyDispatcher==2.0.6
Pygments==2.14.0
PyJWT==2.6.0
pymongo==4.3.2
PyMuPDF==1.21.1
pyOpenSSL==23.0.0
pyparsing==3.0.9
pytest==7.2.0
python-dateutil==2.8.2
python-rtmidi==1.4.9
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.4.1
pywin32==305
PyYAML==6.0
pyzmq==25.0.1
queuelib==1.6.2
regex==2023.3.23
requests==2.28.1
requests-file==1.5.1
requests-oauthlib==1.3.1
rsa==4.9
schedule==1.1.0
scikit-learn==1.1.3
scipy==1.9.3
Scrapy==2.8.0
seaborn==0.12.2
service-identity==21.1.0
simplejson==3.18.0
six==1.16.0
sklearn==0.0.post1
smart-open==6.3.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.44
stack-data==0.6.2
sympy==1.12
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-intel==2.11.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow-probability==0.19.0
termcolor==2.2.0
threadpoolctl==3.1.0
tldextract==3.4.0
tokenizers==0.13.2
toml==0.10.2
tomli==2.0.1
tornado==6.2
tqdm==4.64.1
traitlets==5.9.0
Twisted==22.10.0
twisted-iocpsupport==1.0.2
typing_extensions==4.4.0
tzdata==2022.5
tzlocal==4.2
urllib3==1.26.13
w3lib==2.1.1
wcwidth==0.2.6
webcolors==1.13
Werkzeug==2.2.2
wordcloud==1.8.2.2
wrapt==1.15.0
yarg==0.1.9
zope.interface==5.5.2
